websocket/requirements.txt
Flask>=1.0.3
aio_pika
websockets==6.0
pyyaml
aiohttp==3.5.4
python-socketio
requests>=2.21.0
SQLAlchemy>=1.3.0
psycopg2
pyjwt
